pyproject.toml
[project]
name = "graphisoft-park-restaurant-menu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "David Apagyi", email = "apagyi.david@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "slack-sdk>=3.41.0",
]

[project.scripts]
graphisoft-park-restaurant-menu = "graphisoft_park_restaurant_menu:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
